send If-Modified-Since headers w/ update check
DuckPAN startup is painfully slow, mostly thanks to the numerous HTTP
requests it makes. This change (judging by an entirely unscientific
test) shaves off about 4 seconds from the start. It does so by sending
If-Modified-Since headers with the HTML/JS update checks, based of the
last modified time of the files in cache.

    duckpan server  0.90s user 0.03s system 11% cpu 7.967 total
    duckpan -I../p5-app-duckpan/lib server  0.82s user 0.05s system 24% cpu 3.578 total