DuckPAN/DDG: DuckPAN won't die when a module fails to load.
It's been a perennial problem for DuckPAN users to install all of the dependencies of each IA in the repo before they can even test anything. This patch prevents DuckPAN from exiting when it fails to load an IA. It goes on running, but it warns the user that some IAs didn't load (along with the errors that caused it).