Web.pm: Switch fallback server to ddh1.duckduckgo.com
I would like to change the default fallback server to DDH1 so that people can have access to private APIs.

We sometimes get pull requests from partners that don't have signup pages for APIs. When this happens, only people in DDG have the capability to actually check the API. But when we point it to DDH1, I can deploy the endpoint there and people will be able to play with it on their own DuckPAN installations.

Flow:

1. DuckPAN checks if the env variable is available.
2. If it is, it makes a request to the API endpoint.
3. If it isn't, it makes a request to ddh1.