DuckPAN: add 'reinstall' command.
This is probably a pretty niche need.  If you are doing work on the DDG
or App::DuckPAN modules, you may have installed "newer versions" which
will not necessarily correlate to the version numbers as they will be
published eventually.

`duckpan reinstall` forces duckpan to install the latest version from
the public repostitory without regard to which versions are currently
installed locally.