Server: Unify internal/external name distinction.
We had this concept in a less useful way with the `%assets` hash.
We'd prefer to have them all have similar characteristics so we (and
the code!) can reaosn about what is going on.

With this nearly everything is cached locally, which makes for a better
experiencefor those on slow links. Let's assume those people exist.