App::DuckPAN: move verbose option to App level.
There's no particular reason why verbosity should be set on only the
server command level.  If we move it here, we can better unify logging
and create a more consistent interface.

This is a breaking change for scripts which expect the other ordering,
but it will really improve things.  So let's do it now, early in
product up-take.