DuckPAN: drop line splitting.
It's a lot of wasted cycles in 2014 to do this kind of thing.  If
someone cares, they probably have a terminal setting to do it.  If they
don't, the cycles are even more wasted.

Includes some message review work as well, in an attempt to make sure
this isn't insane.