DuckPAN: bold all prefix lines, NOTE -> NOTICE
The intention is to have them stand out a bit, so this should help.
In the old case, it could blend in with the output if the user happens
to be using a chosen color as their base color, since there aren't
any obvious prefixes on the standard output. If people run with bold all
the time, they deserve what they get.

I dunno why I like NOTICE better, but I somehow do. This means expanding
the prefix_width.. which might be moot now that we don't always prefix.