DuckPAN: improve check_requirements.
This makes it just exit on its own when requirements checking fails.

Also, improve both message text and "emission levels."

Also removes verify_versions which was not well-integrated.  This
involves extra checking on git and SSH which didn't happen to much
before, but is fairly cheap in the grand scheme of things.