check_requirements: move App::DuckPAN check first.
App::DuckPAN itself defines the minimum required and recommended
versions for everything else, so it should be the first thing checked.
Once we get past this check, then all of the other checks will be
properly defined for the installed version.

Also, slight souce reorganization to make it easier to see which things
are being checked and in which order.