Make Web.pm find the correct share dir for IAs that are substring of others
In duckduckgo/zeroclickinfo-spice the Time spice's name is a substring
of the Timer spice's name. When looking for resources for the Timer
IA, paths of the form

  .../timer/...

got mangled to

  .../time/r/...

because the matcher for IAs share directories did not check for the
trailing path separator. Hence, the share path of the Time (no trailing
“r” in Time) matched for Timer (trailing “r” in Timer) resources, and
path handling got confused.

By making the guard for identifying the relevant the share directory
also check on the trailing '/', the share directories of Time and
Timer are properly matched and the Timer IA again properly loads for
'timer 10 seconds'.

Fixes duckduckgo/zeroclickinfo-spice#1625