Use Data::Printer with return_value => 'dump'
A breaking change in Data::Printer 0.36 changed the behavior of p() to always
print instead of returning a string in non-void contexts.

Setting 'return_value' to 'dump' ensures that p() behaves the same as the code
expects it to by printing in a void context and returning a string otherwise.