Adds a hardcoded list of potential author ids to try.
This is a terrible hack.  Production sometimes wants old versions of modules.
duckpan doesn't handle that well because 02packages.details.tar.gz only
contains paths to the latest versions.  Paths are built from the id of the
person who did the release.  This patch helps duckpan cycle through a list of
potential author ids. For each author id it builds a url and checks if the url
points to a valid file on duckpan.org.

A better solution would be to build a service on the server side to resolve
these kinds of queries.  Thats how CPAN works.  I figured this was
easier/faster to build.