Add session caching to Web
Within a single session of say, `duckpan server`, we will only request
external assets such as 'moment.js', 'countries.json', etc. once.

Before this addition I was experiencing delays of between 15 and 20
seconds on each page reload while waiting for assets to be fetched.
After the addition I only experience this on the first page load in a
single session (expected as we need to cache the assets); after that I
experience no noticeable delays.