Switch multiple lookups to an AND operation
Before, when multiple lookups were specified, we would return the union
of the lookups. Now we return the intersection.

This is a bit more useful, as saying something like 'repository Goodies
AND is configured' lets us quickly reduce the possibilities.

Also, it is easier to do `uniq map { my ($by, $lookup) = @$_;
lookup(..., $by, $lookup) } @things...` than to take the intersection (I
may be wrong).

Anyway, for now I think it'll be more useful to return an
'intersection'.