Allow reloading of InstantAnswer::Config files
Files can become outdated, so we need a way of updating them.

It might be better to update them *every* time the files are accessed?
I'm not sure what kind of overhead this would bring...

Maybe a fast way of checking if the IA is up-to-date (a way that is much
faster than actually updating); then we only update the IA if it is
actually outdated.